封装文章列表组件
